backend/requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
axios==0.4.0
bcrypt==4.0.1
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
exceptiongroup==1.2.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
openai==0.28.0
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dotenv==1.0.0
pytz==2022.7.1
redis==5.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
sniffio==1.3.0
spotipy==2.23.0
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==2.3.7
yarl==1.9.3
